get rid of s3 resoure and data source
the huaweicloud_s3_* resource and data source have marked as
deprecated since v1.17.0, use huaweicloud_obs_* instead.
