+++ go.mod
@@ -3,13 +3,10 @@ module github.com/huaweicloud/terraform-provider-huaweicloud
 go 1.14
 
 require (
-	github.com/aws/aws-sdk-go v1.25.3
 	github.com/hashicorp/errwrap v1.0.0
-	github.com/hashicorp/go-cleanhttp v0.5.1
 	github.com/hashicorp/terraform-plugin-sdk v1.13.0
 	github.com/huaweicloud/golangsdk v0.0.0-20210302113304-41351a12edfc
 	github.com/jen20/awspolicyequivalence v0.0.0-20170831201602-3d48364a137a
-	github.com/mitchellh/go-homedir v1.1.0
 	github.com/smartystreets/goconvey v0.0.0-20190222223459-a17d461953aa // indirect
 	github.com/stretchr/testify v1.3.0
 	github.com/unknwon/com v1.0.1
