--- huaweicloud/awserr.go
@@ -1,46 +1,0 @@
-package huaweicloud
-
-import (
-	"time"
-
-	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
-)
-
-func retryOnAwsCode(code string, f func() (interface{}, error)) (interface{}, error) {
-	var resp interface{}
-	err := resource.Retry(1*time.Minute, func() *resource.RetryError {
-		var err error
-		resp, err = f()
-		if err != nil {
-			awsErr, ok := err.(awserr.Error)
-			if ok && awsErr.Code() == code {
-				return resource.RetryableError(err)
-			}
-			return resource.NonRetryableError(err)
-		}
-		return nil
-	})
-	return resp, err
-}
-
-func retryOnAwsCodes(codes []string, f func() (interface{}, error)) (interface{}, error) {
-	var resp interface{}
-	err := resource.Retry(1*time.Minute, func() *resource.RetryError {
-		var err error
-		resp, err = f()
-		if err != nil {
-			awsErr, ok := err.(awserr.Error)
-			if ok {
-				for _, code := range codes {
-					if awsErr.Code() == code {
-						return resource.RetryableError(err)
-					}
-				}
-			}
-			return resource.NonRetryableError(err)
-		}
-		return nil
-	})
-	return resp, err
-}
