+++ huaweicloud/validators.go
@@ -5,7 +5,6 @@ import (
 	"net"
 	"regexp"
 	"strings"
-	"time"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 )
@@ -34,17 +33,6 @@ func ValidateIntRange(v interface{}, k string, l int, h int) (ws []string, error
 	return
 }
 
-func validateS3BucketLifecycleTimestamp(v interface{}, k string) (ws []string, errors []error) {
-	value := v.(string)
-	_, err := time.Parse(time.RFC3339, fmt.Sprintf("%sT00:00:00Z", value))
-	if err != nil {
-		errors = append(errors, fmt.Errorf(
-			"%q cannot be parsed as RFC3339 Timestamp Format", value))
-	}
-
-	return
-}
-
 func validateTrueOnly(v interface{}, k string) (ws []string, errors []error) {
 	if b, ok := v.(bool); ok && b {
 		return
@@ -53,24 +41,6 @@ func validateTrueOnly(v interface{}, k string) (ws []string, errors []error) {
 		return
 	}
 	errors = append(errors, fmt.Errorf("%q must be true", k))
-	return
-}
-
-func validateS3BucketLifecycleExpirationDays(v interface{}, k string) (ws []string, errors []error) {
-	if v.(int) <= 0 {
-		errors = append(errors, fmt.Errorf(
-			"%q must be greater than 0", k))
-	}
-
-	return
-}
-
-func validateS3BucketLifecycleRuleId(v interface{}, k string) (ws []string, errors []error) {
-	value := v.(string)
-	if len(value) > 255 {
-		errors = append(errors, fmt.Errorf(
-			"%q cannot exceed 255 characters", k))
-	}
 	return
 }
